requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
image==1.5.33
iniconfig==2.0.0
MouseInfo==0.1.3
mss==9.0.1
multi-key-dict==2.0.3
numpy==1.26.0
opencv-python==4.8.0.74
outcome==1.2.0
packaging==23.1
pbr==5.11.1
Pillow==10.0.1
pluggy==1.2.0
pyaes==1.6.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pykeyboard==0.1.5
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyperclip==1.8.2
PyRect==0.2.0
Pyrogram==2.0.106
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytest==7.4.0
pytest-assume==2.4.3
pytest-ordering==0.6
python-jenkins==1.8.0
pytweening==1.0.7
PyUserInput==0.1.10
PyYAML==6.0
requests==2.31.0
rubicon-objc==0.4.6
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
TgCrypto==1.2.5
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.7.0
urllib3==2.0.3
wsproto==1.2.0